feat: add local managment fallback

+++ src/store/modules/projects/index.ts
@@ -18,6 +18,8 @@ import {
 } from "@/store/types/actions";
 
 import http from "@/http";
+import { db } from '@/hooks/database';
+import { hasApi } from '@/hooks/verify_api';
 
 export interface ProjectState {
   projects: Project[],
@@ -44,24 +46,52 @@ export const project: Module<ProjectState, State> = {
 
   actions: {
     [FETCH_PROJECTS]: ({ commit }) : Promise<void> => {
-      return http.get('projects')
-        .then(resp => commit(SET_PROJECTS, resp.data));
+      if (hasApi()) {
+        return http.get('projects')
+          .then(resp => commit(SET_PROJECTS, resp.data));
+      }
+
+      return db.collection('projects')
+        .get()
+        .then((resp: Project[]) => commit(SET_PROJECTS, resp));
     },
 
     [CREATE_NEW_PROJECT]: ({ commit }, project_name: string) => {
       const project = { name: project_name };
 
-      return http.post('projects', project)
-        .then(resp => commit(ADD_PROJECT, resp.data));
+      if (hasApi()) {
+        return http.post('projects', project)
+          .then(resp => commit(ADD_PROJECT, resp.data));
+      }
+
+      const new_project = { ...project, id: new Date().getTime() };
+
+      return db.collection('projects')
+        .add(new_project)
+        .then(() => commit(ADD_PROJECT, new_project));
     },
 
     [EDIT_PROJECT]: ({ commit }, project: Project) => {
-      return http.put(`projects/${project.id}`, project)
-        .then(resp => commit(UPDATE_EDITED_PROJECT, resp.data));
+      if (hasApi()) {
+        return http.put(`projects/${project.id}`, project)
+          .then(resp => commit(UPDATE_EDITED_PROJECT, resp.data));
+      }
+
+      return db.collection('projects')
+        .doc({ id: project.id })
+        .update(project)
+        .then(() => commit(UPDATE_EDITED_PROJECT, project));
     },
 
     [DELETE_PROJECT]: ({ commit }, { id: project_id }) => {
-      return http.delete(`projects/${project_id}`)
+      if (hasApi()) {
+        return http.delete(`projects/${project_id}`)
+          .then(() => commit(REMOVE_PROJECT, project_id));
+      }
+
+      return db.collection('projects')
+        .doc({ id: project_id })
+        .delete()
         .then(() => commit(REMOVE_PROJECT, project_id));
     },
   },
